Artefaktor/requirements.txt
# sudo pip3 install -r requirements.txt
Django==1.11.5
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.7.1
social-auth-app-django==2.1.0
djangorestframework-gis==0.12
django-taggit==0.22.2
django-taggit-serializer==0.1.5
